+++ components/eamxx/src/share/pyshare/pyscream.cpp
@@ -1,80 +1,30 @@
 #include "scream_session.hpp"
-#include "share/field/field.hpp"
-#include "share/field/field_utils.hpp"
+#include "pyfield.hpp"
 
 #include <pybind11/pybind11.h>
 #include <pybind11/numpy.h>
 #include <pybind11/stl.h>
 
-using namespace scream;
-using namespace ShortFieldTagsNames;
-namespace py = pybind11;
+namespace scream {
 
 void initialize () {
-  initialize_scream_session(false);
+  initialize_scream_session(true);
 }
 void finalize () {
   finalize_scream_session();
 }
 
-struct PyField {
-  Field f;
-
-  // Create field and allocate memory
-  PyField(const std::string& n,
-          const std::vector<int>& d)
-  {
-    std::vector<FieldTag> t (d.size(),CMP);
-    FieldIdentifier fid(n,FieldLayout(t,d),ekat::units::Units::invalid(),"");
-    f = Field(fid);
-    f.allocate_view();
-  }
-
-  // Create field from pre-existing python nd array
-  PyField(const std::string& n,
-          py::array_t<double> arr)
-  {
-    int rank = arr.ndim();
-    std::vector<int> d(rank,-1);
-    for (int n=0; n<rank; ++n) {
-      d[n] = arr.shape(n);
-    }
-    std::vector<FieldTag> t (rank,CMP);
-    FieldIdentifier fid(n,FieldLayout(t,d),ekat::units::Units::invalid(),"");
-    switch (rank) {
-      case 1:
-      {
-        Field::view_dev_t<double*> v(arr.mutable_data(0),d[0]);
-        f = Field(fid,v);
-        break;
-      }
-      case 2:
-      {
-        Field::view_dev_t<double**> v(arr.mutable_data(0,0),d[0],d[1]);
-        f = Field(fid,v);
-        break;
-      }
-      default:
-        EKAT_ERROR_MSG ("AAARGH!\n");
-    }
-  }
-
-  void print() const {
-    print_field_hyperslab(f);
-  }
-
-  void cleanup () {
-    f = Field();
-  }
-};
+namespace py = pybind11;
 
 PYBIND11_MODULE (pyscream,m) {
 
-  m.doc() = "Basic interface to scream session initialization";
+  m.doc() = "Python interfaces to certain EAMxx infrastructure code";
 
+  // Scream Session
   m.def("init",&initialize);
   m.def("finalize",&finalize);
 
+  // Field class
   py::class_<PyField>(m,"Field")
     .def(py::init<const std::string&,
                   const std::vector<int>&>())
@@ -84,3 +34,5 @@ PYBIND11_MODULE (pyscream,m) {
     .def("print",&PyField::print)
     .def("cleanup",&PyField::cleanup);
 }
+
+} // namespace scream
